// libs/features/collection/billing/src/lib/shared/enums/employer-product-share.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

//FIX these params to be used from API once exposed
export enum EmployerProductShare {
  UI = 1.5,
  ANNUITY = 18,
  OH = 2,
  OH_MOF = 0,
  UI_NEW = 1.5,
  PPA = 18,
  PENSION = 18
}